refactor(usecases/factories): change all affected files to new AccountModel interface
