+++ src/data/usecases/local-save-access-token/local-save-current-account.ts
@@ -1,0 +1,11 @@
+import { type SetStorage } from 'data/protocols/cache'
+import { type AccountModel } from 'domain/models'
+import { type SaveCurrentAccount } from 'domain/usecases'
+
+export class LocalSaveCurrentAccount implements SaveCurrentAccount {
+  constructor(private readonly setStorage: SetStorage) {}
+
+  async save(account: AccountModel): Promise<void> {
+    await this.setStorage.set('currentAccount', JSON.stringify(account))
+  }
+}
