--- src/domain/usecases/save-access-token.ts
@@ -1,3 +1,0 @@
-export interface SaveAccessToken {
-  save: (accessToken: string) => Promise<void> | void
-}
