+++ src/domain/usecases/save-current-account.ts
@@ -1,0 +1,5 @@
+import { type AccountModel } from 'domain/models'
+
+export interface SaveCurrentAccount {
+  save: (accountModel: AccountModel) => Promise<void> | void
+}
